backend/requirements.txt
aiocache==0.11.1
aiocontextvars==0.2.2
aiofiles==0.6.0
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
blinker==1.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
contextvars==2.4
dataclasses==0.8
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.5.4
hyperframe==6.0.0
idna==2.10
idna-ssl==1.1.0
immutables==0.15
itsdangerous==1.1.0
Jinja2==2.11.3
loguru==0.5.3
MarkupSafe==1.1.1
multidict==4.7.6
pycryptodome==3.10.1
pydantic==1.8.1
pytoml==0.1.21
Quart==0.6.15
requests==2.25.1
sortedcontainers==2.3.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wsproto==1.0.0
yarl==1.6.3
redis==3.5.3
python-Levenshtein==0.12.2
gensim==4.0.1
jieba==0.42.1
